Move GlobalStateEvaluator to libsigner and use it in stackerdblistener in miner to retrieve replay transactions
