+++ stacks-signer/src/tests/signer_state.rs
@@ -1,3 +1,17 @@
+// Copyright (C) 2020-2025 Stacks Open Internet Foundation
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 use std::collections::HashMap;
 
 use clarity::types::chainstate::{
@@ -9,287 +23,56 @@ use libsigner::v0::messages::{
     StateMachineUpdate as StateMachineUpdateMessage, StateMachineUpdateContent,
     StateMachineUpdateMinerState,
 };
+use libsigner::v0::signer_state::{GlobalStateEvaluator, SignerStateMachine};
 
 use crate::signerdb::tests::{create_block_override, tmp_db_path};
 use crate::signerdb::SignerDb;
-use crate::v0::signer_state::{GlobalStateEvaluator, SignerStateMachine};
+use crate::v0::signer_state::LocalStateMachine;
 
-fn generate_global_state_evaluator(num_addresses: u32) -> GlobalStateEvaluator {
-    let address_weights = generate_random_address_with_equal_weights(num_addresses);
-    let active_protocol_version = 0;
-    let local_supported_signer_protocol_version = 1;
-
-    let update = StateMachineUpdateMessage::new(
-        active_protocol_version,
-        local_supported_signer_protocol_version,
-        StateMachineUpdateContent::V0 {
-            burn_block: ConsensusHash([0x55; 20]),
-            burn_block_height: 100,
-            current_miner: StateMachineUpdateMinerState::ActiveMiner {
-                current_miner_pkh: Hash160([0xab; 20]),
-                tenure_id: ConsensusHash([0x44; 20]),
-                parent_tenure_id: ConsensusHash([0x22; 20]),
-                parent_tenure_last_block: StacksBlockId([0x33; 32]),
-                parent_tenure_last_block_height: 1,
-            },
-        },
-    )
-    .unwrap();
-
-    let mut address_updates = HashMap::new();
-    for address in address_weights.keys() {
-        address_updates.insert(*address, update.clone());
-    }
-    GlobalStateEvaluator::new(address_updates, address_weights)
-}
-
-fn generate_random_address_with_equal_weights(num_addresses: u32) -> HashMap<StacksAddress, u32> {
+#[test]
+fn check_capitulate_miner_view() {
     let mut address_weights = HashMap::new();
-    for _ in 0..num_addresses {
+    for _ in 0..5 {
         let stacks_address = StacksAddress::p2pkh(
             false,
             &StacksPublicKey::from_private(&StacksPrivateKey::random()),
         );
         address_weights.insert(stacks_address, 10);
     }
-    address_weights
-}
 
-#[test]
-fn determine_latest_supported_signer_protocol_versions() {
-    let mut global_eval = generate_global_state_evaluator(5);
-
-    let addresses: Vec<_> = global_eval.address_weights.keys().cloned().collect();
-    let local_address = addresses[0];
-
-    let local_update = global_eval
-        .address_updates
-        .get(&local_address)
-        .unwrap()
-        .clone();
-    assert_eq!(
-        global_eval
-            .determine_latest_supported_signer_protocol_version(local_address, &local_update,)
-            .unwrap(),
-        local_update.local_supported_signer_protocol_version
-    );
-
-    let StateMachineUpdateMessage {
-        active_signer_protocol_version,
-        local_supported_signer_protocol_version,
-        content:
-            StateMachineUpdateContent::V0 {
-                burn_block,
-                burn_block_height,
-                current_miner,
-            },
-        ..
-    } = local_update.clone()
-    else {
-        panic!("Unexpected state machine update message version");
-    };
-
-    // Let's update 3 signers (60 percent) to support seperate but greater protocol versions
-    for (i, address) in addresses.into_iter().skip(1).take(3).enumerate() {
-        let new_version = local_update.local_supported_signer_protocol_version + i as u64 + 1;
-        let new_update = StateMachineUpdateMessage::new(
-            active_signer_protocol_version,
-            new_version,
-            StateMachineUpdateContent::V0 {
-                burn_block,
-                burn_block_height,
-                current_miner: current_miner.clone(),
-            },
-        )
-        .unwrap();
-        global_eval.insert_update(address, new_update);
-    }
-
-    assert_eq!(
-        global_eval
-            .determine_latest_supported_signer_protocol_version(local_address, &local_update)
-            .unwrap(),
-        local_supported_signer_protocol_version
-    );
-
-    // Let's tip the scales over to version number 2 by updating the local signer's version...
-    // i.e. > 70% will have version 2 or higher in their map
-    let local_update = StateMachineUpdateMessage::new(
-        active_signer_protocol_version,
-        3,
-        StateMachineUpdateContent::V0 {
-            burn_block,
-            burn_block_height,
-            current_miner,
-        },
-    )
-    .unwrap();
-
-    assert_eq!(
-        global_eval
-            .determine_latest_supported_signer_protocol_version(local_address, &local_update)
-            .unwrap(),
-        local_supported_signer_protocol_version + 1
-    );
-}
-
-#[test]
-fn determine_global_burn_views() {
-    let mut global_eval = generate_global_state_evaluator(5);
-
-    let addresses: Vec<_> = global_eval.address_weights.keys().cloned().collect();
-    let local_address = addresses[0];
-    let local_update = global_eval
-        .address_updates
-        .get(&local_address)
-        .unwrap()
-        .clone();
-    let StateMachineUpdateMessage {
-        active_signer_protocol_version,
-        local_supported_signer_protocol_version,
-        content:
-            StateMachineUpdateContent::V0 {
-                burn_block,
-                burn_block_height,
-                current_miner,
-            },
-        ..
-    } = local_update.clone()
-    else {
-        panic!("Unexpected state machine update message version");
-    };
-
-    assert_eq!(
-        global_eval
-            .determine_global_burn_view(local_address, &local_update)
-            .unwrap(),
-        (burn_block, burn_block_height)
-    );
-
-    // Let's update 3 signers (60 percent) to support a new burn block view
-    let new_update = StateMachineUpdateMessage::new(
-        active_signer_protocol_version,
-        local_supported_signer_protocol_version,
-        StateMachineUpdateContent::V0 {
-            burn_block,
-            burn_block_height: burn_block_height.wrapping_add(1),
-            current_miner: current_miner.clone(),
-        },
-    )
-    .unwrap();
-    for address in addresses.into_iter().skip(1).take(3) {
-        global_eval.insert_update(address, new_update.clone());
-    }
-
-    assert!(
-        global_eval
-            .determine_global_burn_view(local_address, &local_update)
-            .is_none(),
-        "We should not have reached agreement on the burn block height"
-    );
-
-    // Let's tip the scales over to burn block height + 1
-    assert_eq!(
-        global_eval
-            .determine_global_burn_view(local_address, &new_update)
-            .unwrap(),
-        (burn_block, burn_block_height.wrapping_add(1))
-    );
-}
-
-#[test]
-fn determine_global_states() {
-    let mut global_eval = generate_global_state_evaluator(5);
-
-    let addresses: Vec<_> = global_eval.address_weights.keys().cloned().collect();
-    let local_address = addresses[0];
-    let local_update = global_eval
-        .address_updates
-        .get(&local_address)
-        .unwrap()
-        .clone();
-    let StateMachineUpdateMessage {
-        active_signer_protocol_version,
-        local_supported_signer_protocol_version,
-        content:
-            StateMachineUpdateContent::V0 {
-                burn_block,
-                burn_block_height,
-                current_miner,
-            },
-        ..
-    } = local_update.clone()
-    else {
-        panic!("Unexpected state machine update message version");
-    };
-
-    let state_machine = SignerStateMachine {
-        burn_block,
-        burn_block_height,
-        current_miner: (&current_miner).into(),
-        active_signer_protocol_version: local_supported_signer_protocol_version, // a majority of signers are saying they support version the same local_supported_signer_protocol_version, so update it here...
-        tx_replay_set: None,
-    };
-
-    assert_eq!(
-        global_eval
-            .determine_global_state(local_address, &local_update)
-            .unwrap(),
-        state_machine
-    );
-    let new_miner = StateMachineUpdateMinerState::ActiveMiner {
-        current_miner_pkh: Hash160([0x00; 20]),
+    let active_signer_protocol_version = 0;
+    let local_supported_signer_protocol_version = 0;
+    let burn_block = ConsensusHash([0x55; 20]);
+    let burn_block_height = 100;
+    let parent_tenure_id = ConsensusHash([0x22; 20]);
+    let parent_tenure_last_block = StacksBlockId([0x33; 32]);
+    let parent_tenure_last_block_height = 1;
+    let old_miner = StateMachineUpdateMinerState::ActiveMiner {
+        current_miner_pkh: Hash160([0xab; 20]),
         tenure_id: ConsensusHash([0x44; 20]),
-        parent_tenure_id: ConsensusHash([0x22; 20]),
-        parent_tenure_last_block: StacksBlockId([0x33; 32]),
-        parent_tenure_last_block_height: 1,
+        parent_tenure_id,
+        parent_tenure_last_block,
+        parent_tenure_last_block_height,
     };
-
-    let new_update = StateMachineUpdateMessage::new(
+    let old_update = StateMachineUpdateMessage::new(
         active_signer_protocol_version,
         local_supported_signer_protocol_version,
         StateMachineUpdateContent::V0 {
             burn_block,
             burn_block_height,
-            current_miner: new_miner.clone(),
+            current_miner: old_miner.clone(),
         },
     )
     .unwrap();
 
-    // Let's update 3 signers to some new miner key (60 percent)
-    for address in addresses.into_iter().skip(1).take(3) {
-        global_eval.insert_update(address, new_update.clone());
+    let mut address_updates = HashMap::new();
+    for address in address_weights.keys() {
+        address_updates.insert(*address, old_update.clone());
     }
-
-    assert!(
-        global_eval
-            .determine_global_state(local_address, &local_update)
-            .is_none(),
-        "We should have a disagreement about the current miner"
-    );
-
-    let state_machine = SignerStateMachine {
-        burn_block,
-        burn_block_height,
-        current_miner: (&new_miner).into(),
-        active_signer_protocol_version: local_supported_signer_protocol_version, // a majority of signers are saying they support version the same local_supported_signer_protocol_version, so update it here...
-        tx_replay_set: None,
-    };
-
-    // Let's tip the scales over to a different miner
-    assert_eq!(
-        global_eval
-            .determine_global_state(local_address, &new_update)
-            .unwrap(),
-        state_machine
-    )
-}
-
-#[test]
-fn check_capitulate_miner_view() {
-    let mut global_eval = generate_global_state_evaluator(5);
+    let mut global_eval = GlobalStateEvaluator::new(address_updates, address_weights);
 
     let addresses: Vec<_> = global_eval.address_weights.keys().cloned().collect();
+    // Let's say we are the very first signer in the list
     let local_address = addresses[0];
     let local_update = global_eval
         .address_updates
@@ -312,14 +95,26 @@ fn check_capitulate_miner_view() {
     };
     // Let's create a new miner view
     let new_tenure_id = ConsensusHash([0x00; 20]);
+
+    let db_path = tmp_db_path();
+    let mut db = SignerDb::new(db_path).expect("Failed to create signer db");
+    let (mut block_info_1, _block_proposal) = create_block_override(|b| {
+        b.block.header.consensus_hash = new_tenure_id;
+        b.block.header.miner_signature = MessageSignature([0x01; 65]);
+        b.block.header.chain_length = 1;
+        b.burn_height = burn_block_height;
+    });
+
+    db.insert_block(&block_info_1).unwrap();
     let new_miner = StateMachineUpdateMinerState::ActiveMiner {
         current_miner_pkh: Hash160([0x00; 20]),
         tenure_id: new_tenure_id,
-        parent_tenure_id: ConsensusHash([0x22; 20]),
-        parent_tenure_last_block: StacksBlockId([0x33; 32]),
-        parent_tenure_last_block_height: 1,
+        parent_tenure_id,
+        parent_tenure_last_block,
+        parent_tenure_last_block_height,
     };
 
+    // Let's update only our own view: the evaluator will tell me to revert my viewpoint to the old miner
     let new_update = StateMachineUpdateMessage::new(
         active_signer_protocol_version,
         local_supported_signer_protocol_version,
@@ -331,20 +126,18 @@ fn check_capitulate_miner_view() {
     )
     .unwrap();
 
-    let db_path = tmp_db_path();
-    let mut db = SignerDb::new(db_path).expect("Failed to create signer db");
-    let (mut block_info_1, _block_proposal) = create_block_override(|b| {
-        b.block.header.consensus_hash = new_tenure_id;
-        b.block.header.miner_signature = MessageSignature([0x01; 65]);
-        b.block.header.chain_length = 1;
-        b.burn_height = 1;
-    });
+    let signer_state_machine = SignerStateMachine {
+        burn_block,
+        burn_block_height,
+        current_miner: (&new_miner).into(),
+        tx_replay_set: None,
+        active_signer_protocol_version,
+    };
 
-    db.insert_block(&block_info_1).unwrap();
-    // Let's update only our own view: the evaluator will tell me to revert my viewpoint to the original miner
+    let mut local_state_machine = LocalStateMachine::Initialized(signer_state_machine.clone());
     assert_eq!(
-        global_eval
-            .capitulate_miner_view(&mut db, local_address, &new_update)
+        local_state_machine
+            .capitulate_miner_view(&mut global_eval, &mut db, local_address, &new_update)
             .unwrap(),
         current_miner
     );
@@ -356,8 +149,8 @@ fn check_capitulate_miner_view() {
         global_eval.insert_update(address, new_update.clone());
     }
     assert!(
-        global_eval
-            .capitulate_miner_view(&mut db, local_address, &new_update)
+        local_state_machine
+            .capitulate_miner_view(&mut global_eval, &mut db, local_address, &new_update)
             .is_none(),
         "Evaluator should have been unable to determine a majority view and return none"
     );
@@ -368,8 +161,8 @@ fn check_capitulate_miner_view() {
 
     // Now that the blocking minority references a tenure which would actually get reorged, lets capitulate to their view
     assert_eq!(
-        global_eval
-            .capitulate_miner_view(&mut db, local_address, &new_update)
+        local_state_machine
+            .capitulate_miner_view(&mut global_eval, &mut db, local_address, &new_update)
             .unwrap(),
         new_miner
     );
