+++ testnet/stacks-node/src/nakamoto_node/stackerdb_listener.rs
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, HashMap, HashSet};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::mpsc::Receiver;
 #[cfg(test)]
@@ -21,21 +21,20 @@ use std::sync::LazyLock;
 use std::sync::{Arc, Condvar, Mutex};
 use std::time::Duration;
 
-use hashbrown::{HashMap, HashSet};
 use libsigner::v0::messages::{
     BlockAccepted, BlockResponse, MessageSlotID, SignerMessage as SignerMessageV0,
-    StateMachineUpdate, StateMachineUpdateContent,
+    StateMachineUpdate,
 };
-use libsigner::{SignerEvent, SignerSession, StackerDBSession};
+use libsigner::v0::signer_state::{GlobalStateEvaluator, SignerStateMachine};
+use libsigner::{SignerEntries, SignerEvent, SignerSession, StackerDBSession};
 use stacks::burnchains::Burnchain;
 use stacks::chainstate::burn::BlockSnapshot;
 use stacks::chainstate::nakamoto::NakamotoBlockHeader;
 use stacks::chainstate::stacks::boot::{NakamotoSignerEntry, RewardSet, SIGNERS_NAME};
 use stacks::chainstate::stacks::events::StackerDBChunksEvent;
-use stacks::chainstate::stacks::{Error as ChainstateError, StacksTransaction};
+use stacks::chainstate::stacks::Error as ChainstateError;
 use stacks::codec::StacksMessageCodec;
-use stacks::net::stackerdb::StackerDBs;
-use stacks::types::chainstate::StacksPublicKey;
+use stacks::types::chainstate::{StacksAddress, StacksPublicKey};
 use stacks::types::PublicKey;
 use stacks::util::get_epoch_time_secs;
 use stacks::util::hash::{MerkleHashFunc, Sha512Trunc256Sum};
@@ -74,12 +73,6 @@ pub(crate) struct TimestampInfo {
     pub weight: u32,
 }
 
-#[derive(Debug, Clone)]
-pub(crate) struct ReplayInfo {
-    pub transactions: Vec<StacksTransaction>,
-    pub weight: u32,
-}
-
 /// The listener for the StackerDB, which listens for messages from the
 /// signers and tracks the state of block signatures and idle timestamps.
 pub struct StackerDBListener {
@@ -108,11 +101,10 @@ pub struct StackerDBListener {
     ///  - key: StacksPublicKey
     ///  - value: TimestampInfo
     pub(crate) signer_idle_timestamps: Arc<Mutex<HashMap<StacksPublicKey, TimestampInfo>>>,
-    /// Tracks any replay transactions from signers to decide when the miner should
-    /// attempt to replay reorged blocks
-    ///  - key: StacksPublicKey
-    ///  - value: Vec<ReplayInfo>
-    pub(crate) replay_info: Arc<Mutex<HashMap<StacksPublicKey, ReplayInfo>>>,
+    /// Tracks the signer's global state machine through signer state machine update messages
+    pub(crate) global_state_evaluator: Arc<Mutex<GlobalStateEvaluator>>,
+    /// Wehther we are operating on mainnet
+    is_mainnet: bool,
 }
 
 /// Interface for other threads to retrieve info from the StackerDBListener
@@ -126,11 +118,8 @@ pub struct StackerDBListenerComms {
     ///  - key: StacksPublicKey
     ///  - value: TimestampInfo
     signer_idle_timestamps: Arc<Mutex<HashMap<StacksPublicKey, TimestampInfo>>>,
-    /// Tracks any replay transactions from signers to decide when the miner should
-    /// attempt to replay reorged blocks
-    ///  - key: StacksPublicKey
-    ///  - value: ReplayInfo
-    replay_info: Arc<Mutex<HashMap<StacksPublicKey, ReplayInfo>>>,
+    /// Tracks the signer's global state machine through signer state machine update messages
+    global_state_evaluator: Arc<Mutex<GlobalStateEvaluator>>,
 }
 
 impl StackerDBListener {
@@ -192,21 +181,17 @@ impl StackerDBListener {
             .ok_or_else(|| ChainstateError::MinerAborted)?;
         let mut signers_session =
             StackerDBSession::new(&rpc_socket.to_string(), signers_contract_id.clone());
-        let stackerdbs = StackerDBs::connect(&config.get_stacker_db_file_path(), false)?;
-        let slot_ids: Vec<_> = stackerdbs
-            .get_signers(&signers_contract_id)
-            .expect("FATAL: could not get signers from stacker DB")
-            .into_iter()
-            .enumerate()
-            .map(|(slot_id, _)| {
-                u32::try_from(slot_id).expect("FATAL: too many signers to fit into u32 range")
-            })
-            .collect();
+        let entries: Vec<_> = signer_entries.values().cloned().collect();
+        let parsed_entries = SignerEntries::parse(config.is_mainnet(), &entries)
+            .expect("FATAL: could not parse retrieved signer entries");
+        let address_weights = parsed_entries.signer_addr_to_weight;
+        let slot_ids: Vec<_> = parsed_entries.signer_id_to_addr.keys().cloned().collect();
+
         let chunks = signers_session
             .get_latest_chunks(&slot_ids)
             .inspect_err(|e| warn!("Unable to read the latest signer state from signer db: {e}."))
             .unwrap_or_default();
-        let mut replay_infos = HashMap::new();
+        let mut global_state_evaluator = GlobalStateEvaluator::new(HashMap::new(), address_weights);
         for (chunk, slot_id) in chunks.into_iter().zip(slot_ids) {
             let Some(chunk) = chunk else {
                 continue;
@@ -217,21 +202,11 @@ impl StackerDBListener {
             let Ok(signer_pubkey) = StacksPublicKey::from_slice(&signer_entry.signing_key) else {
                 continue;
             };
+            let address = StacksAddress::p2pkh(config.is_mainnet(), &signer_pubkey);
             if let Ok(SignerMessageV0::StateMachineUpdate(update)) =
                 SignerMessageV0::consensus_deserialize(&mut chunk.as_slice())
             {
-                let transactions = match update.content {
-                    StateMachineUpdateContent::V0 { .. } => vec![],
-                    StateMachineUpdateContent::V1 {
-                        replay_transactions,
-                        ..
-                    } => replay_transactions,
-                };
-                let replay_info = ReplayInfo {
-                    transactions,
-                    weight: signer_entry.weight,
-                };
-                replay_infos.insert(signer_pubkey, replay_info);
+                global_state_evaluator.insert_update(address, update);
             }
         }
 
@@ -246,7 +221,8 @@ impl StackerDBListener {
             signer_entries,
             blocks: Arc::new((Mutex::new(HashMap::new()), Condvar::new())),
             signer_idle_timestamps: Arc::new(Mutex::new(HashMap::new())),
-            replay_info: Arc::new(Mutex::new(replay_infos)),
+            global_state_evaluator: Arc::new(Mutex::new(global_state_evaluator)),
+            is_mainnet: config.is_mainnet(),
         })
     }
 
@@ -254,7 +230,7 @@ impl StackerDBListener {
         StackerDBListenerComms {
             blocks: self.blocks.clone(),
             signer_idle_timestamps: self.signer_idle_timestamps.clone(),
-            replay_info: self.replay_info.clone(),
+            global_state_evaluator: self.global_state_evaluator.clone(),
         }
     }
 
@@ -522,7 +498,7 @@ impl StackerDBListener {
                         debug!("Received mock message. Ignoring.");
                     }
                     SignerMessageV0::StateMachineUpdate(update) => {
-                        self.update_replay_info(signer_pubkey, signer_entry.weight, update);
+                        self.update_global_state_evaluator(&signer_pubkey, update);
                     }
                 };
             }
@@ -548,30 +524,14 @@ impl StackerDBListener {
         idle_timestamps.insert(signer_pubkey, timestamp_info);
     }
 
-    fn update_replay_info(
-        &self,
-        signer_pubkey: StacksPublicKey,
-        weight: u32,
-        update: StateMachineUpdate,
-    ) {
-        let transactions = match update.content {
-            StateMachineUpdateContent::V0 { .. } => vec![],
-            StateMachineUpdateContent::V1 {
-                replay_transactions,
-                ..
-            } => replay_transactions,
-        };
-        let mut replay_infos = self
-            .replay_info
+    fn update_global_state_evaluator(&self, pubkey: &StacksPublicKey, update: StateMachineUpdate) {
+        let mut eval = self
+            .global_state_evaluator
             .lock()
-            .expect("FATAL: failed to lock idle timestamps");
+            .expect("FATAL: failed to lock global state evaluator");
 
-        // Update the map with the replay info and weight
-        let replay_info = ReplayInfo {
-            transactions,
-            weight,
-        };
-        replay_infos.insert(signer_pubkey, replay_info);
+        let address = StacksAddress::p2pkh(self.is_mainnet, pubkey);
+        eval.insert_update(address, update);
     }
 
     /// Do we ignore signer signatures?
@@ -700,30 +660,12 @@ impl StackerDBListenerComms {
         u64::MAX
     }
 
-    /// Get the transactions that at least 70% of the signing power expect to be replayed in
-    /// the next stacks block
-    pub fn get_replay_transactions(&self, weight_threshold: u32) -> Vec<StacksTransaction> {
-        let replay_info = self
-            .replay_info
+    /// Get the global state if there is one
+    pub fn get_signer_global_state(&self) -> Option<SignerStateMachine> {
+        let mut eval = self
+            .global_state_evaluator
             .lock()
-            .expect("FATAL: failed to lock replay transactions");
-
-        let replay_info = replay_info.values().collect::<Vec<_>>();
-        let mut weights: HashMap<&Vec<StacksTransaction>, u32> = HashMap::new();
-        for info in replay_info {
-            // We only care about signers voting for us to replay a specific set of transactions
-            if info.transactions.is_empty() {
-                continue;
-            }
-            let entry = weights.entry(&info.transactions).or_default();
-            *entry += info.weight;
-            if *entry >= weight_threshold {
-                debug!("SignerCoordinator: 70% threshold reached to attempt replay transactions";
-                    "replay_transactions" => ?info.transactions,
-                );
-                return info.transactions.clone();
-            }
-        }
-        vec![]
+            .expect("FATAL: failed to lock global state evaluator");
+        eval.determine_global_state()
     }
 }
